+++ internal/models/order_easyjson.go
@@ -107,7 +107,7 @@ func easyjson120d1ca2DecodeGithubComFlutterDizasterGophermartBonusInternalModels
 		case "status":
 			out.Status = OrderStatus(in.String())
 		case "accrual":
-			out.Accrual = uint64(in.Uint64())
+			out.Accrual = float64(in.Float64())
 		case "uploaded_at":
 			if data := in.Raw(); in.Ok() {
 				in.AddError((out.UploadedAt).UnmarshalJSON(data))
@@ -139,7 +139,7 @@ func easyjson120d1ca2EncodeGithubComFlutterDizasterGophermartBonusInternalModels
 	if in.Accrual != 0 {
 		const prefix string = ",\"accrual\":"
 		out.RawString(prefix)
-		out.Uint64(uint64(in.Accrual))
+		out.Float64(float64(in.Accrual))
 	}
 	{
 		const prefix string = ",\"uploaded_at\":"
